Fix several compiler warnings and build errors (patch by Daniel Richard G., thanks).
git-svn-id: https://geany.svn.sourceforge.net/svnroot/geany/trunk@3460 ea778897-0a13-0410-b9d1-a72fbfd435f5

+++ src/callbacks.c
@@ -1216,7 +1216,7 @@ on_go_to_line_activate                 (GtkMenuItem     *menuitem,
 
 		if (doc != NULL)
 		{
-			if (! editor_goto_line(doc->editor, val - 1))
+			if (! editor_goto_line(doc->editor, (gint) val - 1))
 				utils_beep();
 		}
 	}
